+++ frontend/src/main.tsx
@@ -1,20 +1,16 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { CssBaseline } from '@mui/material'
-import { ThemeProvider, createTheme } from '@mui/material/styles'
 import { BrowserRouter } from 'react-router-dom'
 import './index.css'
 import App from './App.tsx'
-
-const theme = createTheme()
+import { ThemeModeProvider } from './providers/ThemeModeProvider'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <BrowserRouter>
-      <ThemeProvider theme={theme}>
-        <CssBaseline />
+      <ThemeModeProvider>
         <App />
-      </ThemeProvider>
+      </ThemeModeProvider>
     </BrowserRouter>
   </StrictMode>
 )